restructure funding functions file
